Define new content collection
We'll create a setting in `config.js` to set the path to the posts directory. This directory can be outside of the astro project directory (and is a relative path, based on the astro project root dir). If this is not set in `config.js`, default to `src/posts`.

Define the blog collection, include the elements already used in dummy posts, even if features (e.g. dateModified, redirects, draft) are not yet implemented.

To allow for the same boilerplate frontmatter to be generated in new markdown files, we'll allow `description`, `dateModified`, `image` and `redirects` to be blank/null.

Note: setting `optional` and `nullable` allows the field to be included as blank or not included in frontmatter at all.

If the author is not set in `config.js`, the default author (as far as content collection config is concerned) is set to a blank string.

Draft is set to true by default (i.e. if not set).

`slug` is important. This will be used to set the relative URL(i.e. after `example.com/`) following the move to a content collection. Before moving, the markdown filename (minus the extension) becomes the path. In future, this will allow files in the posts repo to be organised in directories (i.e. year, month) rather than all needing to be in a flat directory structure.