Add tag option check logic
Check whether tag option exists/is an array, and where required, that there are items in the array.

E.g. showing tags under posts on the homepage... a for-loop of 0 items will function, but would still display spacing above the non-existent tags, so we want to check for this eventuality.