Update PostLayout.astro
Okaaay... where to being.

Meta section was above article section, so prose styling (including width) differed. Moved meta into article to simplify styling.

Had a weird bug where the code blocks affected the responsiveness of a page. `grid grid-1` solved this. I think because it's within a flex col. Referenced github issue explains more.

Refactor logic for image display, added a div to contain the image. Div hides overflow, sets height to arbitrary value. Image is set to full width and height, and is vertically centred to equally crop top and bottom. Cover also set to ensure image is zoomed to fit, not stretched.

Prose seemd to be affected when adding it within a flex col (or when adding the grid grid-cols-1).

`<article class="prose mt-2 prose-headings:-my-2 prose-h2:mb-2">` >>

`<article
      class="prose grid grid-cols-1 mt-0 prose-headings:-my-0 prose-h2:mb-2 prose-h2:mt-4 prose-p:mb-1 max-w-2xl p-6"
    >`

Tried adding a config option to set the image height which does get set, and the class is correct in resulting HTML, but Tailwind doesn't generate the CSS for it. Set to static 400px for now.