test SSG `meta http-equiv="refresh"` redirects
Not sure if this is a good solution.

For hosting a static site somewhere like GitHub pages, you don't have access to configure the web server. If migrating content from one blog/URL pattern to another, the only option is to create a small html page at each original URL, with a `meta http-equiv="refresh"` directive to redirect the user.

This can be set in the astro config, but this change also allows it to be set in the frontmatter of each post, on a post by post basis.

The next commit will be the rename of [slug].astro to [...slug].astro, to support redirecting multi/level/paths to path (separate commit for diff reasons).

This meta refresh method isn't the best option if using something like vercel, or a platform with a static adaptor (see: https://docs.astro.build/en/reference/configuration-reference/#redirects), but it's possibly useful for anyone wanting to keep each post's content and config together.

Mixing and matching should also be possible.